pyproject.toml
[project]
name = "pgsocr"
version = "0.1.5"
description = "A command line utility for converting Blu-ray subs to SRT or ASS using AI Language Models."
authors = [
    {name = "pcpeasant", email = "pcpeasant25@gmail.com"},
]
dependencies = [
    "pillow>=10.4.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0,<3.3.0",  # Pin to <3.3.0 due to oneDNN bug in 3.3.0
    "numpy>=2.0.0",
    "tqdm>=4.66.4",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
lm = [
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "torchaudio>=2.3.1",
    "transformers>=4.41.2",
    "einops>=0.8.0",
    "timm>=1.0.7",
    "bitsandbytes>=0.43.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
pgsocr = "pgsocr.cli.main:main"
